Extract library functionality from gradle plugin
Keep only gradle parts in the plugin
